Section Fixes
Fixed the contact section to not be so heavy on performance.

(NEED TO RE DO CONTACT SECTION LATER AGAIN TO MATCH REST OF THE SITE, THIS IS A TEMPORARY CHANGE)

Fixed the nav not properly displaying brands.

Changed Preloader to have a white loading bar instead of the gradient of all companies.